Add files via upload
duplicates matchescombo2 and inserted 4 columns- 14th colum hold a 1 if the match is the results of adding one of the three exact match pairs. then made a dataset that contains the 157 problems that are the result of some questionable situations. found 1 situation where the alignment occurs after the difference in two matches and they are adding the same sequence. the n made function to make the comparisons easier. its called problemCompare and the input is the index of the problem case in matchesComboNoGap2