Add files via upload
added code to fix homozygous mismatch for first time. new data frame called newCOmbinedNoGapAllAdded and has 7021 entries with 5 columns. 1st column is capital sequence with the added istuation in the front and the next two are the name twice and the last two are the same sequence twice in lower caser